Reported by Raymond Penners, inconsistency to ming.h in addString

+++ mingpp.h
@@ -639,7 +639,7 @@ class SWFText : public SWFCharacter
   void setColor(unsigned char r, unsigned char g, unsigned char b, int a=0xff)
     { SWFText_setColor(this->text, r, g, b, a); }
 
-  void addString(const char *string, float *advance=NULL)
+  void addString(const char *string, int *advance=NULL)
     { SWFText_addString(this->text, string, advance); }
 
   void setSpacing(float spacing)
